remove twrnc and fix various things
