--- packages/design-system/tailwind/index.ts
@@ -1,5 +1,0 @@
-import { create } from "twrnc";
-
-const tw = create(require("./tailwind.config.js"));
-
-export { tw };
