+++ packages/design-system/toast/index.tsx
@@ -32,7 +32,7 @@ const ToastContext = createContext<ToastContext | undefined>(undefined);
 const SAFE_AREA_TOP = 20;
 
 export const ToastProvider = ({ children }: any) => {
-  const tw = useTailwind();
+  const tailwind = useTailwind();
   const [show, setShow] = useState(false);
   const [message, setMessage] = useState<string | undefined>();
   const [render, setRender] = useState<React.ReactElement | null>();
@@ -94,7 +94,9 @@ export const ToastProvider = ({ children }: any) => {
               style={[
                 styles.toastContainer,
                 { opacity: layout ? 1 : 0 },
-                tw("bg-white dark:bg-black shadow-black dark:shadow-white"),
+                tailwind(
+                  "bg-white dark:bg-black shadow-black dark:shadow-white"
+                ),
               ]}
               accessibilityLiveRegion="polite"
               pointerEvents="box-none"
